api_integration/requirements.txt
# ==========================
# FRAMEWORK WEB
# ==========================
fastapi==0.111.1
uvicorn[standard]==0.23.2

# ==========================
# VALIDATION / SCHEMAS
# ==========================
pydantic==2.6.2

# ==========================
# CLIENT HTTP ASYNC
# ==========================
httpx==0.28.1

# ==========================
# AUTH / JWT / OAUTH2
# ==========================
python-jose[cryptography]==3.2.0
python-multipart==0.0.7

# ==========================
# ELASTICSEARCH / LOGGING
# ==========================
elasticsearch==8.11.0
loguru==0.7.1

# ==========================
# TESTS / DEV
# ==========================
pytest==8.3.2
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.28.1

# ==========================
# UTILITAIRES
# ==========================
python-dotenv==1.0.0
faker==23.3.0
numpy==1.26.2
requests==2.32.5
